ClienteController implementado
