+++ src/infrastructure/adapters/repositories/interfaces/gerente.repository.interface.ts
@@ -1,5 +1,5 @@
 // src/domain/repositories/gerente.repository.ts
-import { Gerente } from '../../entities/gerente.entity';
+import { Gerente } from '../../../../domain/entities/gerente.entity';
 
 export interface IGerenteRepository {
   findById(id: string): Promise<Gerente | null>;
